refactor(billing): move balance and billing routes
